// src/lib/enums/error.enum.ts
export enum Errors {
  INCORRECT_PASSWORDS = 'Contraseña incorrecta',
  USER_NOT_FOUND = 'Usuario no encontrado',
  ERROR_UPDATE_DATE = 'Error al actualizar la fecha',
  ERROR_SING_IN = "Error al iniciar session",
  USER_EXISTE = 'El usuario ya existe',
  ERROR_SING_UP = 'Error al crear el usuario',
  FAIL_AUTHORIZATION = 'Fallo de autorización ',
  NULL = '',
  EXPIRED_TOKEN = "Token expirado",
  AUTHORIZATION_REQUIRED = "Autorización requerida",
  ERROR_GENERATE_TOKENS = 'Error al generar el tokens'
}